requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
pandas==2.1.0
numpy==1.25.2
matplotlib==3.7.2
plotly==5.16.0
seaborn==0.12.2
openpyxl==3.1.2
python-dotenv==1.0.0
Flask-WTF==1.1.1
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0